+++ end_to_end_test/lib/standard_json.g.dart
@@ -85,6 +85,7 @@ class _$StandardJsonValueSerializer
 // Target: abstract class StandardJsonValue
 // **************************************************************************
 
+// ignore_for_file: annotate_overrides
 class _$StandardJsonValue extends StandardJsonValue {
   @override
   final num number;