+++ example/lib/generics.g.dart
@@ -253,6 +253,7 @@ class _$GenericContainerSerializer
 // Target: abstract class GenericValue
 // **************************************************************************
 
+// ignore_for_file: annotate_overrides
 class _$GenericValue<T> extends GenericValue<T> {
   @override
   final T value;
@@ -337,6 +338,7 @@ class GenericValueBuilder<T>
 // Target: abstract class BoundGenericValue
 // **************************************************************************
 
+// ignore_for_file: annotate_overrides
 class _$BoundGenericValue<T extends num> extends BoundGenericValue<T> {
   @override
   final T value;
@@ -422,6 +424,7 @@ class BoundGenericValueBuilder<T extends num>
 // Target: abstract class CollectionGenericValue
 // **************************************************************************
 
+// ignore_for_file: annotate_overrides
 class _$CollectionGenericValue<T> extends CollectionGenericValue<T> {
   @override
   final BuiltList<T> values;
@@ -511,6 +514,7 @@ class CollectionGenericValueBuilder<T>
 // Target: abstract class GenericContainer
 // **************************************************************************
 
+// ignore_for_file: annotate_overrides
 class _$GenericContainer extends GenericContainer {
   @override
   final GenericValue<String> genericValue;