+++ example/lib/serializers.dart
@@ -5,12 +5,13 @@
 library serializers;
 
 import 'package:built_collection/built_collection.dart';
-import 'package:built_value/serializer.dart';
 import 'package:built_value/json_object.dart';
+import 'package:built_value/serializer.dart';
 import 'package:example/collections.dart';
 import 'package:example/enums.dart';
 import 'package:example/generics.dart';
 import 'package:example/interfaces.dart';
+import 'package:example/polymorphism.dart';
 import 'package:example/values.dart';
 
 part 'serializers.g.dart';